put debug messages with more information back, move checks methods with the respective classes
